add authentication in backend

+++ backend/src/main/java/com/petconnect/backend/entity/Appointment.java
@@ -17,7 +17,7 @@ public class Appointment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer appointmentId;
+    private Long appointmentId;
 
     @Temporal(TemporalType.TIMESTAMP)
     @Column(nullable = false)
